Agent-sync-main/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
azure-ai-inference==1.0.0b9
azure-core>=1.30.0
python-multipart==0.0.6 